requirements.txt
certifi==2020.11.8
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-xlib==0.28
pytz==2020.4
selenium==3.141.0
six==1.15.0
urllib3==1.25.11
